requirements.txt
pylint 
pytest 
pytest-mock 
duolingo-api 
rumps